refactor: v1 제거 및 파일 위치 변경
