+++ src/sockets/utils/auth.ts
@@ -1,5 +1,5 @@
-import { gotClient } from '../../../plugins/http.client.js';
-import { UnAuthorizedException } from '../../../common/exceptions/core.error.js';
+import { gotClient } from '../../plugins/http.client.js';
+import { UnAuthorizedException } from '../../common/exceptions/core.error.js';
 
 export async function verifyAccessToken(token: string): Promise<{
   status: number;
