+++ src/sockets/utils/errorHandler.ts
